+++ src/grandorgue/GODefinitionFile.cpp
@@ -867,6 +867,8 @@ void GODefinitionFile::UpdateTuning() {
     wxString::Format(_("%f cent"), m_PipeConfig.GetPipeConfig().GetTuning()));
 }
 
+void GODefinitionFile::UpdateReleaseTruncationLength() {}
+
 void GODefinitionFile::UpdateAudioGroup() {}
 
 bool GODefinitionFile::IsCustomized() { return m_b_customized; }