+++ src/grandorgue/GODefinitionFile.h
@@ -117,6 +117,7 @@ private:
 
   void UpdateAmplitude();
   void UpdateTuning();
+	void UpdateReleaseTruncationLength();
   void UpdateAudioGroup();
 
   wxString GetOrganHash();