+++ src/grandorgue/GOSoundingPipe.h
@@ -38,6 +38,7 @@ private:
   unsigned m_HarmonicNumber;
   unsigned m_LoopCrossfadeLength;
   unsigned m_ReleaseCrossfadeLength;
+	unsigned m_ReleaseTruncationLength;
   float m_PitchCorrection;
   float m_MinVolume;
   float m_MaxVolume;
@@ -64,6 +65,7 @@ private:
   void SetTremulant(bool on);
 
   void UpdateAmplitude();
+	void UpdateReleaseTruncationLength();
   void UpdateTuning();
   void UpdateAudioGroup();