+++ src/grandorgue/sound/GOSoundEngine.h
@@ -43,6 +43,7 @@ private:
   bool m_RandomizeSpeaking;
   int m_Volume;
   unsigned m_ReleaseLength;
+  unsigned m_ReleaseTruncationLength;
   unsigned m_SamplesPerBuffer;
   float m_Gain;
   unsigned m_SampleRate;
@@ -98,7 +99,12 @@ public:
   int GetVolume() const;
   void SetScaledReleases(bool enable);
   void SetRandomizeSpeaking(bool enable);
-  void SetReleaseLength(unsigned reverb);
+  /* Set Release Length Truncation for Toolbar */
+	void SetReleaseLength(unsigned reverb);
+	/* Get Release Length Truncation Mechanism Value for ODF/Organ Settings Panel */
+	int GetReleaseTruncationLength();
+	/* Set Release Length Truncation Mechanism Value for ODF/Organ Settings Panel */
+	void SetReleaseTruncationLength(unsigned truncation);
   const std::vector<double> &GetMeterInfo();
   void SetAudioRecorder(GOSoundRecorder *recorder, bool downmix);